[#1005] Fix SIG-11 in a SELECT query with a 20-way LEFT JOIN
Background
----------
* This was first noticed as part of https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/985#note_1496159032.

* Below is a simple test case that reproduces the issue with the `names` data set.

  ```m
  $ cat genquery.m
   write "select count(*)",!
   write "from names n0",!
   for i=1:1:$zcmdline do
   . write "left join names n",i," ON n",i,".id = n",(i-1),".id",!
   write ";",!
  ```

* If one uses the above M program to generate a 19-way LEFT JOIN query, it works fine.

  ```sql
  $ yottadb -run genquery 19 | octo
  count(*)
  6
  (1 row)
  ```

* In this case, the generated M program is 781Mb in size.

* But if I try using the above M program to generate a 20-way LEFT JOIN query, it generates a SIG-11.

  ```sql
  $ yottadb -run genquery 20 | octo
  %YDB-F-KILLBYSIGSINFO1, YottaDB process 98750 has been killed by a signal 11 at address 0x00007FE9A9B30C7E (vaddr 0x0000000000000000)
  %YDB-F-SIGMAPERR, Signal was caused by an address not mapped to an object
  ```

* Below are relevant details from the core file.

  ```c
  #5  0x00007fe9aa29225f in generic_signal_handler (sig=11, info=<optimized out>, context=0x7fe9aa585e28 <stapi_signal_handler_oscontext+4552>, is_os_signal_handler=<optimized out>) at /Distrib/YottaDB/R138/sr_unix/generic_signal_handler.c:492
  #6  <signal handler called>
  #7  __memmove_avx_unaligned_erms () at ../sysdeps/x86_64/multiarch/memmove-vec-unaligned-erms.S:749
  #8  0x0000555e46955e4c in resize_tmpl_buffer (global_buffer=0x7fff7e759440, buffer_len=0x7fff7e759424, buffer_index=0x7fff7e7594e0) at src/m_templates/resize_tmpl_buffer.c:19
  #9  0x0000555e46a0d2ad in tmpl_tablejoin_body (global_buffer=0x7fff7e759440, buffer_len=0x7fff7e759424, buffer_index=0x7fff7e7594e0, pplan=0x555e47e35c72, dot_count=21, tableName=0x0, columnName=0x0) at build/src/tmpl_tablejoin_body.ctemplate.c:84
  #10 0x0000555e46a0b807 in tmpl_tablejoin (global_buffer=0x7fff7e759440, buffer_len=0x7fff7e759424, buffer_index=0x7fff7e7594e0, pplan=0x555e47e35c72, tablejoin=0x555e47e342ba, cur_key=21, right_join_second_half=0, dot_count=21, tableName=0x0, colum

  (gdb) f 8
  #8  0x0000555e46955e4c in resize_tmpl_buffer (global_buffer=0x7fff7e759440, buffer_len=0x7fff7e759424, buffer_index=0x7fff7e7594e0) at src/m_templates/resize_tmpl_buffer.c:19
  19              memcpy(tmp, *global_buffer, *buffer_index);

  (gdb) list
  14      #include "template_helpers.h"
  15
  16      void resize_tmpl_buffer(char **global_buffer, int *buffer_len, int *buffer_index) {
  17              *buffer_len *= 2;
  18              char *tmp = calloc(*buffer_len, sizeof(char));
  19              memcpy(tmp, *global_buffer, *buffer_index);
  20              free(*global_buffer);
  21              *global_buffer = tmp;
  22              return;
  23      }

  (gdb) p *buffer_len
  $2 = -2147483648

  (gdb) p/x *buffer_len
  $3 = 0x80000000

  (gdb) p/x *buffer_index
  $4 = 0x3ffffffc
  ```

Issue
-----
* The issue is that `buffer_len` is of type `int` which is a signed 4-byte quantity. It needs to be
  unsigned and an 8-byte quantity to handle generated M code that is more than 2G in size.

Fix
---
* The variables `buffer_len` and `buffer_index` are used in various places in the code with type `int`.
  They have been fixed to instead be `uint64_t`.

* In addition, a local variable inside the `TEMPLATE_SNPRINTF` macro has also been fixed to be `uint64_t`
  as it should be the same type as the above 2 variables.

Notes
-----
* This fixes the SIG-11 while emitting the M code in Octo. But this now generates a huge M program
  that is 1.5Gb in size and compiling this swamps the system as it requires around 58Gb of memory.
  And even then it takes around 4 minutes to finish compilation. Fixing this requires rework to the
  Octo logic that generates M code for LEFT JOINs. It should be changed to not nest `FOR` loops if
  there are too many left joins but instead compute intermediate results of the join. I am not sure
  if that is easily doable at this point. Will try it out and if not easily possible will keep
  #1005 open until that is resolved. That fix, when it happens, will anyways happen in a separate commit.

* Therefore I am not adding a test case for the SIG-11 fix as part of the current commit. Will add an
  automated test case when the LEFT JOIN M code emitting is reworked.

+++ src/m_templates/resize_tmpl_buffer.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- * Copyright (c) 2019-2020 YottaDB LLC and/or its subsidiaries.	*
+ * Copyright (c) 2019-2023 YottaDB LLC and/or its subsidiaries.	*
  * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
@@ -13,7 +13,7 @@
 #include "octo_types.h"
 #include "template_helpers.h"
 
-void resize_tmpl_buffer(char **global_buffer, int *buffer_len, int *buffer_index) {
+void resize_tmpl_buffer(char **global_buffer, uint64_t *buffer_len, uint64_t *buffer_index) {
 	*buffer_len *= 2;
 	char *tmp = calloc(*buffer_len, sizeof(char));
 	memcpy(tmp, *global_buffer, *buffer_index);
